+++ src/netstack/tcp_server.c
@@ -47,7 +47,7 @@ static errval_t server_init(TCP_server* server) {
     // Now we have a new server, we need to start a new thread(s) for it
     for (size_t i = 0; i < server->worker_num; i++) {
         char* name = calloc(16, sizeof(char));
-        sprintf(name, "TCP Server %d", i);
+        sprintf(name, "TCP Server %d", (int)i);
 
         local[i] = (LocalState) {
             .my_name  = name,
